+++ src/main/java/com/eventiming/form2/Controller/PostController.java
@@ -8,8 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
-
 @RestController
 public class PostController {
     @Autowired
@@ -17,8 +15,8 @@ public class PostController {
     @Autowired
     private TokenMangeer tokenMangeer;
     @PostMapping("/post")
-    public int createPost(@Param("userid")BigInteger userid ,
-                          @Param("topicid")BigInteger topicid,
+    public int createPost(@Param("userid")long userid ,
+                          @Param("topicid")long topicid,
                           @Param("postcontext") String postcontext,
                           @Param("token") String token){
         if(tokenMangeer.confirmToken(userid, token))
