+++ src/main/java/com/eventiming/form2/Service/AdminServiceImpl.java
@@ -6,9 +6,6 @@ import com.eventiming.form2.DAO.userdao;
 import com.eventiming.form2.DAO.userstatusDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.math.BigInteger;
-
 @Service
 public class AdminServiceImpl implements AdminService{
     @Autowired
@@ -22,17 +19,17 @@ public class AdminServiceImpl implements AdminService{
 
     @Autowired
     private postDao postdao;
-    public int blockUser(BigInteger userid){
+    public int blockUser(long userid){
         return userstatusdao.updateUserGroupById(userid, 2);
     }
-    public int deleteUser(BigInteger userid){
+    public int deleteUser(long userid){
         userstatusdao.updateUserGroupById(userid,3);
         return userd.deleteUserById(userid);
     }
-    public int deleteTopic(BigInteger topicid){
+    public int deleteTopic(long topicid){
         return topicdao.deleteTopic(topicid);
     }
-    public int deletePost(BigInteger postid){
+    public int deletePost(long postid){
         return postdao.deletePostById(postid);
     }
 }
