+++ src/main/java/com/eventiming/form2/Service/PostService.java
@@ -1,17 +1,17 @@
 package com.eventiming.form2.Service;
 
 import com.eventiming.form2.pojo.post;
-import java.math.BigInteger;
+
 import java.util.List;
 
 public interface PostService {
-    int addpost(BigInteger userid, BigInteger topicid, String postcontext);
+    int addpost(long userid, long topicid, String postcontext);
 
-    int deletepost(BigInteger postid);
+    int deletepost(long postid);
 
     // TODO 支持更改
-    List<post> selectPostByTopic(BigInteger topicid);
+    List<post> selectPostByTopic(long topicid);
 
-    List<post> selectPostByUser(BigInteger userid);
+    List<post> selectPostByUser(long userid);
 
 }
