+++ src/main/java/com/eventiming/form2/pojo/userstatus.java
@@ -4,13 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+
 import java.sql.Timestamp;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class userstatus {
-    BigInteger userid;
+    long userid;
     int usergroup;
     Timestamp lastactivetime;
     int topicnum;
