+++ src/main/java/com/eventiming/form2/util/TokenMangeer.java
@@ -2,20 +2,20 @@ package com.eventiming.form2.util;
 
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import java.math.BigInteger;
+
 import java.sql.Timestamp;
 import java.util.HashMap;
 import java.util.UUID;
 
 @Component
 public class TokenMangeer {
-    private static HashMap<BigInteger, String> onlineToken;
-    private static HashMap<BigInteger, Timestamp> TokenActiveTime;
+    private static HashMap<Long, String> onlineToken;
+    private static HashMap<Long, Timestamp> TokenActiveTime;
     public TokenMangeer(){
         onlineToken =new HashMap<>();
         TokenActiveTime = new HashMap<>();
     }
-    public String getToken(BigInteger userid){
+    public String getToken(long userid){
         if(onlineToken.containsKey(userid)){
             return null;
         }
@@ -26,7 +26,7 @@ public class TokenMangeer {
         return token;
     }
 
-    public boolean confirmToken(BigInteger userid, String token){
+    public boolean confirmToken(long userid, String token){
         if(onlineToken.containsKey(userid)){
             if(onlineToken.get(userid).equals(token)){
                 Timestamp timestamp = new Timestamp(System.currentTimeMillis());
@@ -43,15 +43,15 @@ public class TokenMangeer {
     @Scheduled(fixedDelay = 60000)
     private void clearUnActiveToken(){
         Timestamp now = new Timestamp(System.currentTimeMillis());
-        for (HashMap.Entry<BigInteger, Timestamp> entry : TokenActiveTime.entrySet()) {
-            BigInteger key = entry.getKey();
+        for (HashMap.Entry<Long, Timestamp> entry : TokenActiveTime.entrySet()) {
+            long key = entry.getKey();
             Timestamp value = entry.getValue();
             if((now.getTime() - value.getTime()) > 50000){
                 removeToken(key);
             }
         }
     }
-    public void removeToken(BigInteger userid){
+    public void removeToken(long userid){
         onlineToken.remove(userid);
         TokenActiveTime.remove(userid);
     }
